requirements.txt
python-dotenv==0.19.1
discord.py==1.7.3
riotwatcher==3.0.0
pymysql==1.0.2
pandas==1.3.3
psycopg2==2.9.1
